Summary
Im not sure what I did, ask Nickoli
