+++ src/main/cpp/mechanisms/ActiveIntake/ActiveIntakeStatemgr.cpp
@@ -1,0 +1,194 @@
+//====================================================================================================================================================
+// Copyright 2023 Lake Orion Robotics FIRST Team 302
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
+// to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
+// OR OTHER DEALINGS IN THE SOFTWARE.
+//====================================================================================================================================================
+//==============================================================
+// This file is auto generated by FRCrobotCodeGen302.exe Version $CODE_GENERATOR_VERSION$
+// Changes to this file may cause incorrect behavior and will be lost when
+// the code is regenerated, unless the changes are delimited by:
+//  //========= Hand modified code start section x ========
+//                    Your hand written code goes here
+//	//========= Hand modified code end section x ========
+//==============================================================
+
+#include <map>
+
+// FRC includes
+
+// Team 302 includes
+#include <teleopcontrol/TeleopControl.h>
+#include <auton/PrimitiveParams.h>
+#include <mechanisms/MechanismFactory.h>
+#include <mechanisms/base/StateMgr.h>
+#include <mechanisms/StateStruc.h>
+#include <mechanisms/ActiveIntake/ActiveIntake.h>
+#include <mechanisms/ActiveIntake/ActiveIntakeState.h>
+#include <mechanisms/ActiveIntake/ActiveIntakeStateMgr.h>
+#include <robotstate/RobotState.h>
+#include <robotstate/RobotStateChanges.h>
+#include <utils/logging/Logger.h>
+
+// Third Party Includes
+
+//========= Hand modified code start section 0 ========
+
+//========= Hand modified code end section 0 ========
+
+using namespace std;
+
+ActiveIntakeStateMgr *ActiveIntakeStateMgr::m_instance = nullptr;
+ActiveIntakeStateMgr *ActiveIntakeStateMgr::GetInstance()
+{
+    if (ActiveIntakeStateMgr::m_instance == nullptr)
+    {
+        auto ActiveIntakePtr = MechanismFactory::GetMechanismFactory()->GetActiveIntake();
+        if (ActiveIntakePtr != nullptr)
+        {
+            ActiveIntakeStateMgr::m_instance = new ActiveIntakeStateMgr();
+        }
+    }
+    return ActiveIntakeStateMgr::m_instance;
+}
+
+/// @brief    initialize the state manager, parse the configuration file and create the states.
+ActiveIntakeStateMgr::ActiveIntakeStateMgr() : StateMgr(),
+                                               m_activeIntake(MechanismFactory::GetMechanismFactory()->GetActiveIntake()),
+                                               m_currentState(ACTIVEINTAKE_STATE::OFF),
+                                               m_targetState(ACTIVEINTAKE_STATE::OFF),
+                                               m_prevState(ACTIVEINTAKE_STATE::OFF)
+//========= Hand modified code end section 1 ========
+
+{
+    map<string, StateStruc> stateMap;
+    stateMap["OFF"] = m_off;
+    stateMap["INTAKE"] = m_intake;
+    stateMap["HOLD"] = m_hold;
+    stateMap["PLACE"] = m_place;
+
+    Init(m_activeIntake, stateMap);
+    if (m_activeIntake != nullptr)
+    {
+        m_activeIntake->AddStateMgr(this);
+    }
+
+    //========= Hand modified code start section 2 ========
+    RobotState::GetInstance()->RegisterForStateChanges(this, RobotStateChanges::StateChange::ArmRotateState);
+    //========= Hand modified code end section 2 ========
+}
+
+/// @brief  Get the current Parameter parm value for the state of this mechanism
+/// @param PrimitiveParams* currentParams current set of primitive parameters
+/// @returns int state id - -1 indicates that there is not a state to set
+int ActiveIntakeStateMgr::GetCurrentStateParam(PrimitiveParams *currentParams)
+{
+    return static_cast<int>(currentParams->GetActiveIntakeState());
+}
+
+/// @brief Check if driver inputs or sensors trigger a state transition
+void ActiveIntakeStateMgr::CheckForStateTransition()
+{
+    //========= Hand modified code start section 3 ========
+    if (m_activeIntake != nullptr)
+    {
+        CheckForSensorTransitions();
+        if (m_checkGamePadTransitions)
+        {
+            CheckForGamepadTransitions();
+        }
+
+        if (m_targetState != m_currentState && m_targetState != m_prevState)
+        {
+            SetCurrentState(m_targetState, true);
+            m_prevState = m_targetState;
+            RobotState::GetInstance()->PublishStateChange(RobotStateChanges::ActiveIntakeState, m_targetState);
+        }
+    }
+    else
+    {
+        Logger::GetLogger()->LogData(LOGGER_LEVEL::PRINT, std::string("ActiveIntakeStateMgr"), std::string("have ActiveIntake"), false);
+    }
+    //========= Hand modified code end section 3 ========
+}
+
+//========= Hand modified code start section 4 ========
+/// @brief Check for sensor input to transition
+void ActiveIntakeStateMgr::CheckForSensorTransitions()
+{
+    if (m_activeIntake != nullptr)
+    {
+        // ignore sensor if we are less than 15 degrees above ground
+        if (m_activeIntake->IsGamePiecePresent() && MechanismFactory::GetMechanismFactory()->GetArm()->GetPositionDegrees().to<double>() > m_floorThreshold)
+        {
+            m_targetState = ACTIVEINTAKE_STATE::INTAKE;
+        }
+        if (MechanismFactory::GetMechanismFactory()->GetArm()->GetPositionDegrees().to<double>() < m_protectThreshold)
+        {
+            m_targetState = ACTIVEINTAKE_STATE::INTAKE;
+        }
+    }
+    else
+    {
+        Logger::GetLogger()->LogData(LOGGER_LEVEL::PRINT, std::string("ActiveIntakeStateMgr"), std::string("ActiveIntake"), false);
+    }
+}
+
+/// @brief Check for gamepad input to transition
+void ActiveIntakeStateMgr::CheckForGamepadTransitions()
+{
+    if (m_activeIntake != nullptr)
+    {
+        m_currentState = static_cast<ACTIVEINTAKE_STATE>(GetCurrentState());
+        // m_targetState = m_currentState;
+
+        auto controller = TeleopControl::GetInstance();
+        if (controller != nullptr)
+        {
+            if (controller->IsButtonPressed(TeleopControlFunctions::OFF))
+            {
+                m_targetState = ACTIVEINTAKE_STATE::OFF;
+            }
+            else if (controller->IsButtonPressed(TeleopControlFunctions::INTAKE))
+            {
+                m_targetState = ACTIVEINTAKE_STATE::INTAKE;
+            }
+            if (controller->IsButtonPressed(TeleopControlFunctions::HOLD))
+            {
+                m_targetState = ACTIVEINTAKE_STATE::HOLD;
+            }
+            else if (controller->IsButtonPressed(TeleopControlFunctions::PLACE))
+            {
+                m_targetState = ACTIVEINTAKE_STATE::PLACE;
+            }
+        }
+    }
+}
+
+void ActiveIntakeStateMgr::Update(RobotStateChanges::StateChange change, int value)
+{
+    if (change == RobotStateChanges::StateChange::ArmRotateState)
+    {
+        // If we are going to the HP substation, open the ActiveIntake, otherwise follow the sensors or gamepad input
+        // don't update m_followOtherMechs in HOLD_POSITION because we will go to that after reaching HP target, then open the ActiveIntake <- We dont want this
+        ArmStateMgr::ARM_STATE armState = static_cast<ArmStateMgr::ARM_STATE>(value);
+        if (armState == ArmStateMgr::ARM_STATE::HUMAN_PLAYER_STATION_ROTATE)
+        {
+            m_targetState = ACTIVEINTAKE_STATE::OFF;
+            m_followOtherMechs = true;
+        }
+        else if (armState != ArmStateMgr::ARM_STATE::HOLD_POSITION_ROTATE)
+        {
+            m_followOtherMechs = false;
+        }
+    }
+}
+//========= Hand modified code end section 4 ========
