+++ src/main/deploy/states/activeIntake.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<statedata xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <controlData
+    identifier="PercentOutput"
+    mode="PERCENT_OUTPUT"
+    constrolServer="MOTORCONTROLLER"
+    proportional="0.0"
+    integral="0.0"
+    derivative="0.0"
+    feedforward="0.0"
+    izone="0.0"
+    maxacceleration="0.0"
+    cruisevelocity="0.0" />
+
+  <mechanismTarget
+    stateIdentifier="OFF"
+    controlDataIdentifier="PercentOutput"
+    controlDataIdentifier2=""
+    value="0"
+    secondValue=""/>
+    
+  <mechanismTarget
+    stateIdentifier="INTAKE"
+    controlDataIdentifier="PercentOutput"
+    controlDataIdentifier2=""
+    value="1"
+    secondValue="" />
+
+  <mechanismTarget
+    stateIdentifier="HOLD"
+    controlDataIdentifier="PercentOutput"
+    controlDataIdentifier2=""
+    value="0.1"
+    secondValue=""/>
+    
+  <mechanismTarget
+    stateIdentifier="PLACE"
+    controlDataIdentifier="PercentOutput"
+    controlDataIdentifier2=""
+    value="-1"
+    secondValue=""/>
+    
+</statedata>
